requirements.txt
fastapi~=0.115.12
uvicorn
sqlalchemy~=2.0.43
pymysql
cryptography
httpx~=0.28.1
jose~=1.0.0
config~=0.5.1
python-dotenv~=1.2.1
chromadb~=1.3.5
langchain-openai~=1.1.0
pydantic~=2.12.4
langchain-core~=1.1.0
langchain-community~=0.3.30
passlib~=1.7.4